requirements.txt
build
black
PyYAML>=6.0,<7
mypy
